Rename `cfg.rs` file in utils to `config.rs`
