+++ src/utils/mod.rs
@@ -1,2 +1,2 @@
-pub mod cfg;
+pub mod config;
 pub mod paths;
